src/requirements.txt

exa-py
python-dotenv
google-generativeai
requests
supabase